requirements.txt
anyio==3.6.2
asyncpg==0.27.0
bcrypt==4.0.1
certifi==2022.12.7
click==8.1.3
colorama==0.4.6
databases==0.6.1
dnspython==2.2.1
email-validator==1.3.0
fastapi==0.88.0
greenlet==2.0.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.1
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.1
passlib==1.7.4
psycopg2==2.9.5
pydantic==1.10.2
PyJWT==2.6.0
python-decouple==3.6
python-dotenv==0.21.0
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
